requirements.txt
boto3==1.9.1
botocore==1.12.1
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
pkg-resources==0.0.0
python-dateutil==2.7.3
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
termcolor==1.1.0
urllib3==1.23
